% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.r
\docType{package}
\name{extra}
\alias{extra}
\title{Package to compute the log likelihood for an inverse gamma stochastic volatility model and to draw from the exact posterior of the inverse volatilities.}
\description{
This package computes the log likelihood for an inverse gamma stochastic volatility model using a closed form expression of the likelihood.
}
