% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ourgeo}
\alias{ourgeo}
\title{Computes the 2F1 Hypergeometric Function}
\usage{
ourgeo(a1,a2,b1,zstar,niter=500)
}
\arguments{
\item{a1}{Parameter (Real)}

\item{a2}{Parameter (Real)}

\item{b1}{Parameter (Real)}

\item{zstar}{Primary real argument}

\item{niter}{The degree of approximation to truncate the hypergeometric sum. The default value is set at 500}
}
\value{
returns the value of the hypergeometric function
}
\description{
Computes the 2F1 Hypergeometric Function
}
\examples{
##usage of ourgeo to evaluate a 2F1 hypergeometric function
ourgeo(1.5,1.9,1.2,0.7)
}
