% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_Inf_Data}
\alias{US_Inf_Data}
\title{Data to use in the invgamstochvol package}
\format{
{A data frame with 1 variable and 247 observations.}
\describe{
\item{V1}{Inflation variable}
}
}
\source{
{Federal Reserve Bank of St Louis Fred database as the Consumer Price Index (CPI) data to serve as example data}
}
\usage{
US_Inf_Data
}
\description{
Contains the inflation data for the US for the period 1960Q1 to 2022Q3
}
\examples{
data(US_Inf_Data) #lazy loading
}
\keyword{datasets}
